requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.5
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
idna==3.1
multidict==5.1.0
pycodestyle==2.6.0
python-dotenv==0.15.0
pytz==2021.1
toml==0.10.2
typing-extensions==3.7.4.3
yarl==1.6.3
